+++ bundle/src/Entity/OidcSession.php
@@ -1,8 +1,0 @@
-<?php
-
-namespace Hyvor\Internal\Bundle\Entity;
-
-class OidcSession
-{
-
-}