+++ tests/SymfonyTestBundle.php
@@ -2,6 +2,10 @@
 
 namespace Hyvor\Internal\Tests;
 
+use Doctrine\Bundle\DoctrineBundle\Registry;
+use Doctrine\ORM\EntityManager;
+use Doctrine\ORM\EntityManagerInterface;
+use Doctrine\Persistence\ManagerRegistry;
 use Symfony\Component\Cache\Adapter\ArrayAdapter;
 use Symfony\Component\DependencyInjection\ContainerBuilder;
 use Symfony\Component\DependencyInjection\Loader\Configurator\ContainerConfigurator;
@@ -23,6 +27,8 @@ class SymfonyTestBundle extends AbstractBundle
             ->set('kernel', SymfonyKernel::class)
             ->set(HttpClientInterface::class, MockHttpClient::class)
             ->set(CacheInterface::class, ArrayAdapter::class)
+            ->set(EntityManagerInterface::class, EntityManager::class)
+            ->set(ManagerRegistry::class, Registry::class)
             ->public();
     }